Create cli with serve, lint, init ...
