+++ packages/snage/src/command/common.ts
@@ -1,0 +1,3 @@
+export interface DefaultCli {
+    config: string;
+}
