+++ packages/snage/webpack.dev.js
@@ -3,6 +3,6 @@ const WebpackShellPlugin = require('webpack-shell-plugin');
 
 module.exports = {
     ...common,
-    plugins: [...common.plugins, new WebpackShellPlugin({onBuildEnd: ['nodemon build/server.js --watch build']})],
+    plugins: [...common.plugins, new WebpackShellPlugin({onBuildEnd: ['nodemon build/server.js serve --watch build']})],
     mode: 'development',
 };
